backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai==1.51.0
python-dotenv==1.0.1
aiofiles==23.2.1
python-pptx==1.0.2
python-docx==1.1.2
pdfplumber==0.11.4
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.9
